requirements.txt
openai==1.78.0
python-dotenv==1.1.0